添加引导界面的文字提示，修改回档的bug：数据库中删除了相关的数据，管理类没有将其还原，到时后面只进行更新，表中没有这个数据更新失败没法保存

+++ Assets/Scripe/SkillPakeger/AttackSkill/EventAttactSkill/AttackSkill600007.cs
@@ -12,7 +12,7 @@ public class AttackSkill600007 : EventAttackSkill
         if(count1 == 0) { 
             count1 = mSkillJson.getSpecialParameterValue()[0] /100;
         }
-        mManager.getAttacker().AddBlood((int)(count1 * mManager.getAttacker().mAttribute.aggressivity));
+        mManager.getAttacker().AddBlood(count1 * mManager.getAttacker().mAttribute.aggressivity);
         Debug.Log("===============60007 add " + count1 * mManager.getAttacker().mAttribute.aggressivity);
     }
     public override void endSkill()
